requirements.txt
llama-index-core
llama-index-llms-anthropic
llama-index-embeddings-openai
anthropic
